# Using numba more in MultiLine