This gives a 33% speedup in set_msparams:
```
# before
| calc_110748, maxmem=15.2 GB | time_sec | memory_mb | counts |
|-----------------------------+----------+-----------+--------|
| total preclassical          | 3_503    | 34.5      | 297    |
| setting msparams            | 3_318    | 0.0       | 208    |
| PreClassicalCalculator.run  | 386.2    | 2_998     | 1      |

# after
| calc_110746, maxmem=16.1 GB | time_sec | memory_mb | counts |
|-----------------------------+----------+-----------+--------|
| total preclassical          | 2_353    | 60.6      | 297    |
| setting msparams            | 2_178    | 0.0       | 208    |
| PreClassicalCalculator.run  | 317.7    | 2_962     | 1      |
```